Refactor Life implementation for better perf
Restructures the Conways Game of Life swift implementation to make the
hotloop faster. This version matches the C versions performance while
using nicer wrapper types.

Using -Osize causes the binary to be smaller than the C version at the
cost of perf. This commit changes the build to use -O instead, which
causes the binary size to match C, but also matches the performance of
the C version.