Make Sprite: ~Copyable (#13)
Now that support for `UnsafeMutablePointer<Pointee: ~Copyable>` has
landed in the standard library (gated under the
`-enable-experimental-feature NoncopyableGenerics`), we can use it to
simplify the `Sprite` type.

This converts the `Sprite` wrapper for the Playdate SDK from an
enum+class to track ownership to a non-copyable struct that can have a
`deinit` and needs no reference counting. It also adds a very minimal
array type to hold the bricks (since the regular `Array` type cannot yet
hold non-copyable types).

This drops binary size by about 14%. Half of this is from simplifying
the sprite type, and half is from the simpler array type.